Squashed commit of the following:
commit 9e961907e10180845a40095fff438dd28e748e9b
Author: Matt Clarke <matt.clarke@esss.se>
Date:   Fri Apr 16 15:32:52 2021 +0200

    mob next [ci-skip]

commit 713e8a5fa347163fa40a73b96ef31e0e14a39d98
Author: Matthew D Jones <matthew.d.jones@tessella.com>
Date:   Fri Apr 16 14:16:58 2021 +0100

    mob next [ci-skip]

commit 4a581311513743059b07599776e10071bcb4863a
Author: Matt Clarke <matt.clarke@esss.se>
Date:   Fri Apr 16 15:06:11 2021 +0200

    mob next [ci-skip]

commit 49b481997fcce04565c6c5a3d16d6e2509f21642
Author: Matt Clarke <matt.clarke@esss.se>
Date:   Fri Apr 16 14:58:09 2021 +0200

    Include resetting gross when finishing

commit 3bda05e192cca5f4f75a38095bc54493d88d51c1
Author: Matthew D Jones <matthew.d.jones@tessella.com>
Date:   Fri Apr 16 13:55:35 2021 +0100

    mob next [ci-skip]

commit 1f01beddf69aaea80fa88a9e8a029c57bd09514d
Author: Matt Clarke <matt.clarke@esss.se>
Date:   Fri Apr 16 14:45:06 2021 +0200

    mob next [ci-skip]

+++ tests/test_apply_tax.py
@@ -11,7 +11,7 @@ def test_taxable_items_have_tax_applied():
     cart = Cart(catalogue)
     cart.add_item_by_barcode(barcode)
 
-    assert cart.gross_total() == price * TAX_MULTIPLIER
+    assert cart.gross_total() == int(price * TAX_MULTIPLIER)
 
 
 def test_non_taxable_items_do_not_have_tax_applied():
@@ -38,9 +38,9 @@ def test_multiple_items_some_taxed_some_not_taxed():
         catalogue.add_new_product(barcode, price, taxable)
         cart.add_item_by_barcode(barcode, quantity)
 
-    contribution_from_item_1 = 1234 * TAX_MULTIPLIER
-    contribution_from_item_2 = 3273 * 3
-    contribution_from_item_3 = 234 * TAX_MULTIPLIER * 2
+    contribution_from_item_1 = int(1234 * TAX_MULTIPLIER)
+    contribution_from_item_2 = int(3273 * 3)
+    contribution_from_item_3 = int(234 * TAX_MULTIPLIER * 2)
     assert (
         cart.gross_total()
         == contribution_from_item_1
